Add -Weverything compilation flag and files update

+++ srcs/env_treat.c
@@ -6,7 +6,7 @@
 /*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/16 03:31:29 by nmougino          #+#    #+#             */
-/*   Updated: 2016/09/17 00:09:20 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/09/17 18:27:19 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ static t_env	*env_new(char *name, char *cont)
 char	*env_extract(char **env, char *tar)
 {
 	char	*ans;
-	int		len;
+	size_t	len;
 
 	len = ft_strlen(tar);
 	while (ft_strncmp(*env, tar, len))
